% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Spacial plot of fatal accidents per state - Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer value, id state.}

\item{year}{integer value, reference year.}
}
\value{
This function a map plot with the pattern points of injuries for state - year
}
\description{
This function generate a plot with the pattern points of fatal accidents that happened in a state during a year
}
\details{
The files "accident_\code{year}.csv.bz2" contain the American public data regarding fatal injuries suffered
in motor vehicle traffic crashes in a especified \code{year}. The \code{fars_map_state} function read the individual files asociated to \code{year} and generate a spacial map with the patterns points for a state especified by user.

The file "accident_\bold{year}.csv.bz2" should be download to working director
}
\note{
An error message is generated when the file "accident_\bold{year}.csv.bz2" isn't found in the working directory.
#' The accidents locate in \eqn{Lattitude <90} or \eqn{Longitud > 900} are omitted.
}
\examples{
\dontrun{
fars_map_state(4,2013)
fars_map_state(8,2015)
}

}
