% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV file as a tibble object}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character vector, containing filename or path for the CSV-file which the \code{fars_read} function will search.}
}
\value{
This function returns the original CSV-file as a tibble object.
        For the examples, the data ("accident_\bold{year}.csv.bz2") should be download to working directry from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{ FARS}
}
\description{
This function search a CSV file and read it as a tibble object.
}
\details{
The tibble objects are an efficient structure to storage and handling a \code{data.frame}.
  See more details about tibble objects in CRAN. \url{https://cran.r-project.org/web/packages/tibble/vignettes/tibble.html}
}
\note{
The \emph{"simple"} diagnostics messages and reading progress bar are omitted.
      An error message is generated when the file isn't found.
}
\examples{
\dontrun{
 fars_read ("accident_2013.csv.bz2")
 fars_read ("accident_2014.csv.bz2")
 fars_read ("accident_2015.csv.bz2")
}
}
