% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{List of Month/Year accidents by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Integers vector which contain the years to include in the list.}
}
\value{
This function return a list of tibble object with the Month and Year variables
}
\description{
This function create a list of tibble object with the Month and Year associated with the report of an injury. Each item on the list corresponds to one year and each register corresponds to an injury report for this year.
}
\details{
The files "accident_\code{year}.csv.bz2" contain the American public data regarding fatal injuries suffered
in motor vehicle traffic crashes in a especified \code{year}. The \code{fars_read_years} function extract the Month and Year Variable from individual files asociated to \code{years} and locate it into list

The files "accident_\bold{year}.csv.bz2" should be download to working directory from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System. \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}
}
\note{
When the individual file associated to i-th year isn't found, the function generates a warnings message and the i-th element on list will be NULL
}
\examples{
\dontrun{
fars_read_years(c(2013,2014,2015))
}

}
