% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Total number of fatal accidents per Month/Year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integers vector which contain the years to include in the summary table.}
}
\value{
This function return a summary table with number of fatal accidents per Month (rows) - Year (columns).
}
\description{
This function generate a summary of the total number of fatal accidents grouped by Month/Year
}
\details{
The files "accident_\code{year}.csv.bz2" contain the American public data regarding fatal injuries suffered
in motor vehicle traffic crashes in a especified \code{year}. The \code{fars_summarize_years} function read the individual files asociated to \code{years} and generate a summary table with the number of injuries for Month -Year combination.

The files "accident_\bold{year}.csv.bz2" should be download to working directory from the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System. \href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{FARS}
}
\note{
An error message is generated when the file "accident_\bold{year}.csv.bz2" isn't found in the working directory.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014,2015))
}
}
