% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Adding a year label to the filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Integer value, define the year to include in the filename.}
}
\value{
Character string with the filename/year merge
}
\description{
This function generate a new filename wich merge the original filename "accident.csv.bz2 " with a reference year especified by user.
}
\details{
The files "accident_\bold{year}.csv.bz2" contain the American public data regarding fatal injuries suffered
in motor vehicle traffic crashes in a especified \code{year}. The \code{make_filename} function merge the \code{year} with the root "accident.csv.bz2" to obtain the true filename associated to the specified \code{year} .
         This function is required by \code{\link{fars_read_years}} functions.
}
\examples{

make_filename(2013)
make_filename(2014)
make_filename(2015)

}
